+++ src/alarmUtils.ts
@@ -1,6 +1,6 @@
 import sdk, { SecuritySystemMode } from "@scrypted/sdk";
 import { StorageSetting, StorageSettings } from "@scrypted/sdk/storage-settings";
-import { ExtendedNotificationAction, getWebhooks, sensorsFilter } from "./utils";
+import { ExtendedNotificationAction, getAssetsParams, getWebhooks, sensorsFilter } from "./utils";
 import AdvancedNotifierPlugin from "./main";
 
 export const supportedAlarmModes = [
@@ -190,16 +190,11 @@ export const getAlarmWebhookUrls = async (props: {
     const { setAlarm } = await getWebhooks();
 
     try {
-        const cloudEndpointRaw = await sdk.endpointManager.getCloudEndpoint(undefined, { public: false });
-        const cloudEndpointUrl = new URL(cloudEndpointRaw);
-        const cloudEndpoint = cloudEndpointUrl.origin;
-        const userToken = cloudEndpointUrl.searchParams.get('user_token');
-
-        const { privateKey } = plugin.storageSettings.values;
-        let paramString = `?secret=${privateKey}`;
-        if (userToken) {
-            paramString += `&user_token=${userToken}`;
-        }
+        const {
+            paramString,
+            publicPathnamePrefix,
+            assetsEndpoint,
+        } = await getAssetsParams({ plugin });
 
         for (const alarmMode of [
             SecuritySystemMode.Disarmed,
@@ -217,7 +212,7 @@ export const getAlarmWebhookUrls = async (props: {
                 title = setModeMessage.replace('${mode}', modeText);
             }
             actions.push({
-                url: `${cloudEndpoint}${publicPathnamePrefix}${setAlarm}/${alarmMode}${paramString}`,
+                url: `${assetsEndpoint}${publicPathnamePrefix}${setAlarm}/${alarmMode}&${paramString}`,
                 title,
                 action: `scrypted_an_alarm_${alarmMode}`,
             });