+++ src/utils.ts
@@ -302,6 +302,44 @@ export const isDetectionRule = (rule: BaseRule) => [
     RuleType.Detection,
 ].includes(rule.ruleType);
 
+export const getAssetsParams = async (props: {
+    plugin: AdvancedNotifierPlugin,
+}) => {
+    try {
+        const { plugin } = props;
+        const { serveAssetsFromLocal } = plugin.storageSettings.values;
+        const localEndpoint = await sdk.endpointManager.getLocalEndpoint(undefined, { public: true });
+        const cloudEndpoint = await sdk.endpointManager.getLocalEndpoint(undefined, { public: true });
+
+        const endpointRaw = serveAssetsFromLocal || !plugin.hasCloudPlugin ? localEndpoint : cloudEndpoint;
+        const endpointUrl = new URL(endpointRaw);
+        const assetsEndpoint = endpointUrl.origin;
+        const userToken = endpointUrl.searchParams.get('user_token');
+
+        const { privateKey } = plugin.storageSettings.values;
+        const searchParams = new URLSearchParams();
+        searchParams.set('secret', privateKey);
+        if (userToken) {
+            searchParams.set('user_token', userToken);
+        }
+        const paramString = `${searchParams.toString()}`;
+
+        const privatePathname = await endpointManager.getPath(undefined, { public: false });
+        const publicPathnamePrefix = await endpointManager.getPath(undefined, { public: true });
+
+        return {
+            paramString,
+            assetsEndpoint,
+            localEndpoint,
+            cloudEndpoint,
+            privatePathname,
+            publicPathnamePrefix
+        };
+    } catch {
+        return {}
+    }
+}
+
 export const getWebHookUrls = async (props: {
     cameraIdOrAction?: string,
     console?: Console,
@@ -333,7 +371,6 @@ export const getWebHookUrls = async (props: {
     let eventImageUrl: string;
     let eventVideoclipUrl: string;
     let privatePathnamePrefix: string;
-    let publicPathnamePrefix: string;
 
     const snoozeActions: ExtendedNotificationAction[] = [];
 
@@ -352,36 +389,25 @@ export const getWebHookUrls = async (props: {
     } = await getWebhooks();
 
     try {
-        let privatePathname: string;
-        let localEndpoint: string;
-        try {
-            localEndpoint = await sdk.endpointManager.getLocalEndpoint(undefined, { public: true });
-            privatePathname = await endpointManager.getPath(undefined, { public: false });
-            publicPathnamePrefix = await endpointManager.getPath(undefined, { public: true });
-        } catch { }
-
-        const cloudEndpointRaw = await sdk.endpointManager.getCloudEndpoint(undefined, { public: false });
-        const cloudEndpointUrl = new URL(cloudEndpointRaw);
-        const cloudEndpoint = cloudEndpointUrl.origin;
-        const userToken = cloudEndpointUrl.searchParams.get('user_token');
-
-        const { privateKey } = plugin.storageSettings.values;
         const encodedId = encodeURIComponent(cameraIdOrAction ?? device?.id);
-        const searchParams = new URLSearchParams();
-        searchParams.set('secret', privateKey);
-        if (userToken) {
-            searchParams.set('user_token', userToken);
-        }
-        const paramString = `&${searchParams.toString()}`;
 
-        lastSnapshotCloudUrl = `${cloudEndpoint}${publicPathnamePrefix}${lastSnapshot}/${encodedId}/{IMAGE_NAME}${paramString}`;
-        lastSnapshotLocalUrl = `${localEndpoint}${lastSnapshot}/${encodedId}/{IMAGE_NAME}${paramString}`;
-        haActionUrl = `${cloudEndpoint}${publicPathnamePrefix}${haAction}/${encodedId}${paramString}`;
-        postNotificationUrl = `${cloudEndpoint}${publicPathnamePrefix}${postNotification}/${encodedId}${paramString}`;
+        const {
+            assetsEndpoint,
+            cloudEndpoint,
+            paramString,
+            localEndpoint,
+            privatePathname,
+            publicPathnamePrefix,
+        } = await getAssetsParams({ plugin });
 
-        videoclipStreamUrl = `${cloudEndpoint}${publicPathnamePrefix}${videoclipStream}/${fileId}${paramString}`;
-        gifUrl = `${cloudEndpoint}${publicPathnamePrefix}${gif}/${fileId}${paramString}`;
-        videoclipThumbnailUrl = `${cloudEndpoint}${publicPathnamePrefix}${videoclipThumbnail}/${fileId}${paramString}`;
+        lastSnapshotCloudUrl = `${cloudEndpoint}${publicPathnamePrefix}${lastSnapshot}/${encodedId}/{IMAGE_NAME}&${paramString}`;
+        lastSnapshotLocalUrl = `${localEndpoint}${lastSnapshot}/${encodedId}/{IMAGE_NAME}&${paramString}`;
+        haActionUrl = `${assetsEndpoint}${publicPathnamePrefix}${haAction}/${encodedId}&${paramString}`;
+        postNotificationUrl = `${assetsEndpoint}${publicPathnamePrefix}${postNotification}/${encodedId}&${paramString}`;
+
+        videoclipStreamUrl = `${assetsEndpoint}${publicPathnamePrefix}${videoclipStream}/${fileId}&${paramString}`;
+        gifUrl = `${assetsEndpoint}${publicPathnamePrefix}${gif}/${fileId}&${paramString}`;
+        videoclipThumbnailUrl = `${assetsEndpoint}${publicPathnamePrefix}${videoclipThumbnail}/${fileId}&${paramString}`;
 
         privatePathnamePrefix = `${privatePathname}${eventsApp}`;
         eventThumbnailUrl = `${privatePathnamePrefix}/${eventThumbnail}/${device?.id}/${fileId}`;
@@ -393,7 +419,7 @@ export const getWebHookUrls = async (props: {
                 const { minutes, text } = snooze;
 
                 snoozeActions.push({
-                    url: `${cloudEndpoint}${publicPathnamePrefix}${snoozeNotification}/${encodedId}/${snoozeId}/${minutes}${paramString}`,
+                    url: `${assetsEndpoint}${publicPathnamePrefix}${snoozeNotification}/${encodedId}/${snoozeId}/${minutes}&${paramString}`,
                     title: text,
                     action: `snooze${minutes}`,
                     data: minutes,
@@ -417,7 +443,6 @@ export const getWebHookUrls = async (props: {
         eventImageUrl,
         eventVideoclipUrl,
         privatePathnamePrefix,
-        publicPathnamePrefix,
         gifUrl
     };
 }